// app/types/disciplineTypes.ts
export enum DisciplineDataType {
    Time = "time",
    Distance = "distance",
    Points = "points",
}

export enum Discipline {
    SixtyMeters = "60m",
    OneHundredMeters = "100m",
    TwoHundredMeters = "200m",
    FourHundredMeters = "400m",
    EightHundredMeters = "800m",
    OneAndAHalfKilometers = "1500m",
    ThreeKilometers = "3000m",
    FiveKilometers = "5000m",
    TenKilometers = "10000m",
    SixtyMetersHurdles = "60m_hurdles",
    OneHundredMetersHurdles = "100m_hurdles",
    OneHundredAndTenMetersHurdles = "110m_hurdles",
    FourHundredMetersHurdles = "400m_hurdles",
    ThreeKilometersSteeplechase = "3000m_steeplechase",
    OneHundredMetersRelay = "4x100m_relay",
    FourHundredMetersRelay = "4x400m_relay",
    HalfMarathon = "half_marathon",
    Marathon = "marathon",
    TwentyKilometersRaceWalk = "20km_race_walk",
    FiftyKilometersRaceWalk = "50km_race_walk",
    CrosCountryRunning = "cross_country_running",
    HighJump = "high_jump",
    LongJump = "long_jump",
    TripleJump = "triple_jump",
    PoleVault = "pole_vault",
    ShotPut = "shot_put",
    Discus = "discus",
    Hammer = "hammer",
    Javelin = "javelin",
    Pentathlon = "pentathlon",
    Heptathlon = "heptathlon",
    Decathlon = "decathlon",
}